# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EventShooter API - Premium event photography booking platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: role
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban a user account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanInput"
      responses:
        "200":
          description: Banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/photographers/{id}/approve:
    post:
      operationId: approvePhotographer
      tags: [admin]
      summary: Approve photographer profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/photographers/{id}/feature:
    post:
      operationId: featurePhotographer
      tags: [admin]
      summary: Feature a photographer on the homepage
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureInput"
      responses:
        "200":
          description: Featured status set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/reviews/{id}/moderate:
    post:
      operationId: moderateReview
      tags: [admin]
      summary: Moderate (approve/remove) a review
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerateInput"
      responses:
        "200":
          description: Moderated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/coupons:
    get:
      operationId: adminListCoupons
      tags: [admin]
      summary: Admin - list all coupons
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: adminCreateCoupon
      tags: [admin]
      summary: Admin - create coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /admin/revenue:
    get:
      operationId: getRevenueStats
      tags: [admin]
      summary: Admin revenue analytics
      parameters:
        - name: period
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Revenue stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueStats"

  # ─── AI FEATURES (stubs) ───
  /ai/cost-estimate:
    post:
      operationId: aiCostEstimate
      tags: [ai]
      summary: AI cost estimator for event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiEstimateInput"
      responses:
        "200":
          description: AI estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiEstimateResponse"
  /ai/package-recommendations:
    post:
      operationId: aiPackageRecommendations
      tags: [ai]
      summary: AI package recommendations based on event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiRecommendInput"
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
  /ai/caption-suggestions:
    post:
      operationId: aiCaptionSuggestions
      tags: [ai]
      summary: AI Instagram caption suggestions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiCaptionInput"
      responses:
        "200":
          description: Captions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiCaptionResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

    # ─── AUTH SCHEMAS ───
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        role: { type: string, enum: [customer, photographer, admin] }
        isVerified: { type: boolean }
        isBanned: { type: boolean }
        city: { type: ["string", "null"] }
        referralCode: { type: ["string", "null"] }
        createdAt: { type: string }
...
    Package:
      type: object
      required: [id, photographerId, name, price, duration]
      properties:
        id: { type: integer }
        photographerId: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        duration: { type: integer }
        includes: { type: array, items: { type: string } }
        addons: { type: array, items: { $ref: "#/components/schemas/PackageAddon" } }
        eventTypes: { type: array, items: { type: string } }
        isActive: { type: boolean }
    PackageAddon:
      type: object
      required: [id, name, price]
      properties:
        id: { type: integer }
        name: { type: string }
        price: { type: number }
        description: { type: ["string", "null"] }
...
    MonthlyEarning:
      type: object
      required: [month, amount, bookings]
      properties:
        month: { type: string }
        amount: { type: number }
        bookings: { type: integer }
...
    Coupon:
      type: object
      required: [id, code, discountType, discountValue, isActive, createdAt]
      properties:
        id: { type: integer }
        code: { type: string }
        discountType: { type: string, enum: [percentage, flat] }
        discountValue: { type: number }
        minOrderAmount: { type: ["number", "null"] }
        maxDiscountAmount: { type: ["number", "null"] }
        usageLimit: { type: ["integer", "null"] }
        usedCount: { type: integer }
        expiresAt: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code: { type: string }
        discountType: { type: string, enum: [percentage, flat] }
        discountValue: { type: number }
        minOrderAmount: { type: number }
        maxDiscountAmount: { type: number }
        usageLimit: { type: integer }
        expiresAt: { type: string }
...
    BanInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    FeatureInput:
      type: object
      required: [isFeatured]
      properties:
        isFeatured: { type: boolean }
    ModerateInput:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [approve, remove] }
    RevenueStats:
      type: object
      required: [totalRevenue, platformFee, netRevenue]
      properties:
        totalRevenue: { type: number }
        platformFee: { type: number }
        netRevenue: { type: number }
        byEventType: { type: array, items: { $ref: "#/components/schemas/RevenueByCategory" } }
        byCity: { type: array, items: { $ref: "#/components/schemas/RevenueByCategory" } }
        monthly: { type: array, items: { $ref: "#/components/schemas/MonthlyEarning" } }
    RevenueByCategory:
      type: object
      required: [category, amount, bookings]
      properties:
        category: { type: string }
        amount: { type: number }
        bookings: { type: integer }

    # ─── AI SCHEMAS ───
    AiEstimateInput:
      type: object
      required: [eventType, city, guestCount, durationHours]
      properties:
        eventType: { type: string }
        city: { type: string }
        guestCount: { type: integer }
        durationHours: { type: number }
        additionalRequirements: { type: string }
    AiEstimateResponse:
      type: object
      required: [minEstimate, maxEstimate, averageEstimate, explanation]
      properties:
        minEstimate: { type: number }
        maxEstimate: { type: number }
        averageEstimate: { type: number }
        explanation: { type: string }
        recommendedPackageType: { type: string }
    AiRecommendInput:
      type: object
      required: [eventType, budget, city]
      properties:
        eventType: { type: string }
        budget: { type: number }
        city: { type: string }
        guestCount: { type: integer }
    AiCaptionInput:
      type: object
      required: [eventType, mood]
      properties:
        eventType: { type: string }
        mood: { type: string }
        keywords: { type: array, items: { type: string } }
    AiCaptionResponse:
      type: object
      required: [captions]
      properties:
        captions: { type: array, items: { type: string } }